requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.7
django-autoslug==1.9.9
idna==3.4
lxml==4.9.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.7
